refactor(ceremony): decompose 2067-line ceremony-service.ts into 5 focused classes
